refactor: 더 보기 버튼 컴포넌트 분리
